Revert to std::size_t in some cases

+++ include/oneapi/dpl/pstl/hetero/dpcpp/parallel_backend_sycl_histogram.h
@@ -421,9 +421,9 @@ struct __histogram_general_private_global_atomics_submitter<__internal::__option
         const std::uint64_t __max_segments =
             std::min(__global_mem_size / (__num_bins * sizeof(_bin_type)), __max_groups);
 
-        const std::uint64_t __iters_per_work_item =
+        const std::size_t __iters_per_work_item =
             oneapi::dpl::__internal::__dpl_ceiling_div(__n, __max_segments * __work_group_size);
-        const std::uint64_t __segments =
+        const std::size_t __segments =
             oneapi::dpl::__internal::__dpl_ceiling_div(__n, __work_group_size * __iters_per_work_item);
 
         auto __private_histograms =
